refactor: Relocate api content
